tools/streaming-ui/requirements.txt
chainlit
gradio
python-dotenv
requests
streamlit
